Authorization implemented

+++ app/Http/Controllers/PostsController.php
@@ -4,22 +4,29 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
-use Illuminate\Database\Eloquent\Model;
+
+use App\Models\User;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
+
 
 class PostsController extends Controller
 {
-        public function __construct()
+    public function __construct()
     {
         $this->middleware('auth')
-            ->only(['create','store','edit', 'update', 'destroy']);
+            ->only(['create', 'store', 'edit', 'update', 'destroy']);
     }
 
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
+     * @return Application|Factory|View|Response
      */
     public function index()
     {
@@ -40,24 +47,25 @@ class PostsController extends Controller
 
         //comments_count
 
-        return view('posts.index', ['posts' => BlogPost::withCount('comments')->get()]);
+        return view('posts.index', ['posts' => BlogPost::withCount('comments')->with('user')->get()]);
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
+     * @return Application|Factory|View|Response
      */
     public function create()
     {
+//        $this->authorize('posts.create');
         return view('posts.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\RedirectResponse
+     * @param Request $request
+     * @return RedirectResponse
      */
     public function store(StorePost $request)
     {
@@ -79,7 +87,7 @@ class PostsController extends Controller
      * Display the specified resource.
      *
      * @param int $id
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
+     * @return Application|Factory|View|Response
      */
     public function show($id)
     {
@@ -93,41 +101,63 @@ class PostsController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param int $id
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
+     * @return Application|Factory|View|Response
      */
     public function edit($id)
     {
-        return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+
+        $post = BlogPost::findorfail($id);
+//        if (Gate::denies('update-post', $post){
+//            abort(403, "You can't update this blog post!")
+//        });
+        $this->authorize($post);// <- zamiast Gate::denies
+
+        return view('update', ['post' => BlogPost::findOrFail($id)]);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
+     * @param Request $request
      * @param int $id
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response
+     * @return RedirectResponse|Response
      */
     public function update(StorePost $request, $id)
     {
-        $post = BlogPost::findOrFail($id);
+
+        $post = BlogPost::findorfail($id);
+
+//        if (Gate::denies('update-post', $post){
+//            abort(403, "You can't update this blog post!")
+//        });
+
+        $this->authorize($post); // <- zamiast Gate::denies
+
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
 
         $request->session()->flash('status', 'Blog post was updated!');
 
-        return redirect()->route('posts.show', ['post' => $post ->id]);
+        return redirect()->route('posts.show', ['post' => $post->id]);
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param int $id
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response
+     * @return RedirectResponse|Response
      */
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+//        if (Gate::denies('delete-post', $post){
+//        abort(403, "You can't delete this blog post!")
+//        }) ;
+        $this->authorize($post);// <- zamiast Gate::denies
+
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
